Update the postFX shader in the draw function
Currently the bump map is updated in the `update` function instead of
the draw function. This makes it possible to observe visual glitches
at certain points because the game is still drawn using the shader,
but with an out-of-date bump map. It also causes a glitch on startup
because the bump map is not initialized.

Fix this by always updating the bump map before drawing. This has the
side effect that heat updates when the game is paused - I think it
looks nice though :)