Set the shaders viewport on every frame
When shaders are enabled, the background is drawn on the window at a
slight offset (depending on the window size).

This happens because `startDrawHUD` resets the viewport to the correct
value for drawing to the window, but it's never set back again. The
`resized` function does set the viewport to the correct value, but
this isn't helpful (it will be reset on the second frame).

Fix by setting the viewport on every frame before using the back
buffer.